+++ Assets/Scripts/Player.cs
@@ -1,10 +1,10 @@
 [System.Serializable]
 public class Player
 {
-    public int ID;
+    public byte ID;
     public bool myTurn;
 
-    public Player(int ID)
+    public Player(byte ID)
     {
         if (ID < 0 || ID > 4)
             throw new System.Exception("Incorrect  player ID");
